+++ tests/test_cli.py
@@ -6,7 +6,7 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 def test_cli_import():
     try:
-        import cli
+        import cli_ui
         import_test = True
     except ImportError:
         import_test = False